[Jobs details] Missing "View YAML" action
