+++ src/components/Table/TableView.js
@@ -20,6 +20,7 @@ import {
   MODELS_PAGE,
   DATASETS_TAB
 } from '../../constants'
+import { ACTIONS_MENU } from '../../types'
 
 const TableView = ({
   actionsMenu,
@@ -255,10 +256,7 @@ TableView.defaultProps = {
 }
 
 TableView.propTypes = {
-  actionsMenu: PropTypes.oneOfType([
-    PropTypes.arrayOf(PropTypes.shape({})),
-    PropTypes.func
-  ]).isRequired,
+  actionsMenu: ACTIONS_MENU.isRequired,
   applyDetailsChanges: PropTypes.func,
   content: PropTypes.arrayOf(PropTypes.shape({})).isRequired,
   getCloseDetailsLink: PropTypes.func,
