+++ src/elements/ProjectCard/ProjectCard.js
@@ -6,6 +6,7 @@ import ProjectCardView from './ProjectCardView'
 
 import { generateProjectStatistic } from './projectCard.util'
 import projectsAction from '../../actions/projects'
+import { ACTIONS_MENU } from '../../types'
 
 const ProjectCard = ({
   actionsMenu,
@@ -57,7 +58,7 @@ const ProjectCard = ({
 }
 
 ProjectCard.propTypes = {
-  actionsMenu: PropTypes.shape({}).isRequired,
+  actionsMenu: ACTIONS_MENU.isRequired,
   project: PropTypes.shape({}).isRequired
 }
