+++ src/elements/ProjectCard/ProjectCardView.js
@@ -7,6 +7,8 @@ import Tooltip from '../../common/Tooltip/Tooltip'
 import TextTooltipTemplate from '../TooltipTemplate/TextTooltipTemplate'
 import ProjectStatistics from '../ProjectStatistics/ProjectStatistics'
 
+import { ACTIONS_MENU } from '../../types'
+
 import './projectCard.scss'
 
 const ProjectCardView = React.forwardRef(
@@ -57,7 +59,7 @@ const ProjectCardView = React.forwardRef(
 )
 
 ProjectCardView.propTypes = {
-  actionsMenu: PropTypes.shape({}).isRequired,
+  actionsMenu: ACTIONS_MENU.isRequired,
   project: PropTypes.shape({}).isRequired,
   statistics: PropTypes.shape({}).isRequired
 }
